combined fileUtil
